Split Filter Action/Object/Tag struct into multiple structs
Generalize the action and tags to be templated on the FilterType.
Move the new structs from Evolution/DiscontinuousGalerkin to
NumericalAlgorithms/LinearOperators
Enable filtering in the 2D scalar wave executable so its use is tested.
