+++ tests/Unit/NumericalAlgorithms/LinearOperators/CMakeLists.txt
@@ -8,6 +8,7 @@ set(LIBRARY_SOURCES
   Test_CoefficientTransforms.cpp
   Test_DefiniteIntegral.cpp
   Test_Divergence.cpp
+  Test_Filtering.cpp
   Test_IndefiniteIntegral.cpp
   Test_Linearize.cpp
   Test_MeanValue.cpp
@@ -19,5 +20,5 @@ add_test_library(
   ${LIBRARY}
   "NumericalAlgorithms/LinearOperators/"
   "${LIBRARY_SOURCES}"
-  "LinearOperators;MathFunctions;Utilities"
+  "LinearOperators;MathFunctions;Spectral;Utilities"
   )
